fix: log dialogs
